# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Palais Divin Backend
  version: 0.0.1-SNAPSHOT
servers: []
paths:
  /api/v1/admin/invitations:
    post:
      operationId: issue
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvitationResponse"
          description: OK
      tags:
      - invitation-rest-controller
  /api/v1/public/ping:
    get:
      operationId: ping
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PingResponse"
          description: OK
      tags:
      - ping-controller
  /api/v1/public/restaurants:
    get:
      operationId: list_2
      parameters:
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: sort
        required: false
        schema:
          type: string
          default: CREATED_AT_DESC
          enum:
          - CREATED_AT_DESC
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantsPageResponse"
          description: OK
      tags:
      - public-restaurant-rest-controller
  /api/v1/public/restaurants/{id}:
    get:
      operationId: get_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantResponse"
          description: OK
      tags:
      - public-restaurant-rest-controller
  /api/v1/public/restaurants/{restaurantId}/reviews:
    get:
      operationId: list_3
      parameters:
      - in: path
        name: restaurantId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: sort
        required: false
        schema:
          type: string
          default: CREATED_AT_DESC
          enum:
          - CREATED_AT_DESC
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewsPageResponse"
          description: OK
      tags:
      - public-review-rest-controller
  /api/v1/public/signup:
    post:
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SignupResponse"
          description: OK
      tags:
      - signup-rest-controller
  /api/v1/user/connections/{targetId}:
    post:
      operationId: connect
      parameters:
      - in: path
        name: targetId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
          description: OK
      tags:
      - connection-rest-controller
  /api/v1/user/recommendations:
    get:
      operationId: list_1
      parameters:
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: sort
        required: false
        schema:
          type: string
          default: AFFINITY_DESC
          enum:
          - AFFINITY_DESC
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecommendationsPageResponse"
          description: OK
      tags:
      - recommendation-rest-controller
  /api/v1/user/restaurants:
    get:
      operationId: list
      parameters:
      - in: query
        name: cursor
        required: false
        schema:
          type: string
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: sort
        required: false
        schema:
          type: string
          default: CREATED_AT_DESC
          enum:
          - CREATED_AT_DESC
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantsPageResponse"
          description: OK
      tags:
      - restaurant-rest-controller
    post:
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantResponse"
          description: OK
      tags:
      - restaurant-rest-controller
  /api/v1/user/restaurants/{id}:
    get:
      operationId: get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantResponse"
          description: OK
      tags:
      - restaurant-rest-controller
  /api/v1/user/restaurants/{restaurantId}/reviews:
    post:
      operationId: create_1
      parameters:
      - in: path
        name: restaurantId
        required: true
        schema:
          type: string
          format: uuid
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
          description: OK
      tags:
      - review-rest-controller
components:
  schemas:
    ConnectionResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        sourceUserId:
          type: string
          format: uuid
        targetUserId:
          type: string
          format: uuid
    CoordinatesDto:
      type: object
      properties:
        latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
      required:
      - latitude
      - longitude
    CreateRestaurantRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
      required:
      - address
      - name
    CreateReviewRequest:
      type: object
      properties:
        comment:
          type: string
          maxLength: 1000
          minLength: 0
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
    InvitationResponse:
      type: object
      properties:
        expiresAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        signupUrl:
          type: string
    PageMeta:
      type: object
      properties:
        hasNext:
          type: boolean
        nextCursor:
          type: string
        size:
          type: integer
          format: int32
    PingResponse:
      type: object
      properties:
        status:
          type: string
        ts:
          type: string
          format: date-time
    RecommendationResponse:
      type: object
      properties:
        address:
          type: string
        affinity:
          type: number
          format: double
        id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        name:
          type: string
        recommenderCount:
          type: integer
          format: int32
    RecommendationsPageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationResponse"
        page:
          $ref: "#/components/schemas/PageMeta"
    RestaurantResponse:
      type: object
      properties:
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        location:
          $ref: "#/components/schemas/CoordinatesDto"
        name:
          type: string
    RestaurantsPageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RestaurantResponse"
        page:
          $ref: "#/components/schemas/PageMeta"
    ReviewResponse:
      type: object
      properties:
        authorId:
          type: string
          format: uuid
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        rating:
          type: integer
          format: int32
        restaurantId:
          type: string
          format: uuid
    ReviewsPageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResponse"
        page:
          $ref: "#/components/schemas/PageMeta"
    SignupRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
      required:
      - displayName
      - email
      - password
      - token
    SignupResponse:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        displayName:
          type: string
        email:
          type: string
        id:
          type: string
          format: uuid