# src/main/java/fr/lepgu/palaisdivin/backend/review/adapters/rest/ReviewsPageResponse.java
package fr.lepgu.palaisdivin.backend.review.adapters.rest;

import fr.lepgu.palaisdivin.backend.shared.adapters.web.PageMeta;
import java.util.List;

public record ReviewsPageResponse(List<ReviewResponse> data, PageMeta page) {}